Add centralized SQL insert helper
